+++ djangosecure/conf.py
@@ -20,6 +20,7 @@ class Configuration(object):
 conf = Configuration(
     SECURE_HSTS_SECONDS=0,
     SECURE_FRAME_DENY=False,
+    SECURE_CONTENT_TYPE_NOSNIFF=False,
     SECURE_SSL_REDIRECT=False,
     SECURE_SSL_HOST=None,
     SECURE_REDIRECT_EXEMPT=[],
@@ -31,6 +32,7 @@ conf = Configuration(
         "djangosecure.check.djangosecure.check_security_middleware",
         "djangosecure.check.djangosecure.check_sts",
         "djangosecure.check.djangosecure.check_frame_deny",
+        "djangosecure.check.djangosecure.check_content_type_nosniff",
         "djangosecure.check.djangosecure.check_ssl_redirect",
         ]
     )