+++ djangosecure/middleware.py
@@ -9,6 +9,7 @@ class SecurityMiddleware(object):
     def __init__(self):
         self.sts_seconds = conf.SECURE_HSTS_SECONDS
         self.frame_deny = conf.SECURE_FRAME_DENY
+        self.content_type_nosniff = conf.SECURE_CONTENT_TYPE_NOSNIFF
         self.redirect = conf.SECURE_SSL_REDIRECT
         self.redirect_host = conf.SECURE_SSL_HOST
         self.proxy_ssl_header = conf.SECURE_PROXY_SSL_HEADER
@@ -44,4 +45,8 @@ class SecurityMiddleware(object):
                 not 'strict-transport-security' in response):
             response["strict-transport-security"] = ("max-age=%s"
                                                      % self.sts_seconds)
+        if (self.content_type_nosniff and
+                not 'x-content-type-options' in response):
+            response["x-content-type-options"] = "nosniff"
+
         return response