+++ CHANGES.rst
@@ -7,6 +7,8 @@ tip (unreleased)
 * ``SECURE_PROXY_SSL_HEADER`` setting now patches ``request.is_secure()`` so it
   respects proxied SSL, to avoid redirects to http that should be to https.
 
+* Added the ``X-Content-Type-Options: nosniff`` header.
+
 
 0.1.0 (2011.05.29)
 ------------------